requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.1.5
django-bootstrap4==2.3.1
django-tempus-dominus==5.1.2.13
docopt==0.6.2
docutils==0.16
idna==3.1
kiwisolver==1.3.1
matplotlib==3.3.3
num2words==0.5.10
numpy==1.20.0rc2
pdfkit==0.6.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pyparsing==3.0.0b2
PyPDF4==1.27.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
uWSGI==2.0.19.1
